requirements.txt
flask==2.2.5
opencv-python
pillow
numpy
